requirements.txt
pandas
psycopg2-binary==2.9.9
sqlalchemy==2.0.32
pymongo==4.6.1
python-dotenv==1.0.0
coverage